added floating numbers

+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -9,9 +9,9 @@ public final class AccountData {
     private final String name;
     private final String email;
 
-    private final int balance;
+    private final float balance;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -30,12 +30,14 @@ public final class AccountData {
         return email;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return balance;
     }
 
     @Override
     public String toString() {
+
+
         return "Account id: " + id + '\n' +
                 "Name: " + name + '\n' +
                 "Email: " + email + '\n' +
