+++ src/main/java/rocks/zipcode/atm/bank/Bank.java
@@ -14,19 +14,19 @@ public class Bank {
 
     public Bank() {
         accounts.put(1000, new BasicAccount(new AccountData(
-                1000, "Dolio Durant", "dolio@gmail.com", 5000
+                1000, "Dolio Durant", "dolio@gmail.com", 5000F
         )));
 
         accounts.put(2000, new PremiumAccount(new AccountData(
-                2000, "Leon Hunter", "leon@gmail.com", 200
+                2000, "Leon Hunter", "leon@gmail.com", 200F
         )));
 
         accounts.put(3000, new PremiumAccount(new AccountData(
-                3000, "Rogelio Gamboa Jr", "jr@gmail.com", 100
+                3000, "Rogelio Gamboa Jr", "jr@gmail.com", 100F
         )));
 
         accounts.put(4000, new PremiumAccount(new AccountData(
-                4000, "Kris Younger", "kris@gmail.com", 50000
+                4000, "Kris Younger", "kris@gmail.com", 50000F
         )));
     }
 
@@ -40,14 +40,14 @@ public class Bank {
         }
     }
 
-    public ActionResult<AccountData> deposit(AccountData accountData, int amount) {
+    public ActionResult<AccountData> deposit(AccountData accountData, float amount) {
         Account account = accounts.get(accountData.getId());
         account.deposit(amount);
 
         return ActionResult.success(account.getAccountData());
     }
 
-    public ActionResult<AccountData> withdraw(AccountData accountData, int amount) {
+    public ActionResult<AccountData> withdraw(AccountData accountData, float amount) {
         Account account = accounts.get(accountData.getId());
         boolean ok = account.withdraw(amount);
